Fix memcpy bug in test_packet_output_ipv6_to_gre
The original code intends to copy the first 64 bits of a 128 bit
IPv6 address into a route msg.

Previously, a temporary variable was used but only the first 16 bits of
the 128 bit address was copied via an assignment stmt. This fix removes
that use of a temporary variable.

This bug was exposed when compiling with CPPFLAGS='-O0'.

Signed-off-by: Brian Brooks <brian.brooks@arm.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>