Fix packet buffer leak in ofp_soreceive_generic()
This patch fixes a packet buffer leak when a user calls e.g. ofp_recv
to read data from a socket. As the data is passed to the user's
buffer, the sb structure used to manage the mbuf chain is updated,
but the backing ODP packet buffer is never released.

The original issue manifested, when running the webserver example, as
an ENOBUF after processing 500-600 TCP requests.

Signed-off-by: Brian Brooks <brian.brooks@arm.com>
Reviewed-by: Jose Pekkarinen <jose.pekkarinen@nokia.com>