Fix OFP_DEBUG compile-time define
This patch fixes the situation where OFP_DEBUG_PRINT was always
passed at compile-time regardless of whether --enable-debug was
specified during the configuration stage.

With this patch, -DOFP_DEBUG will be passed iff --enable-debug
is specified. The OFP_DBG(..) logging macros will now be
compiled-out as intended.

Additionally, any code that is determined to be too much overhead
for a hot code path can be compiled out for a non-debug build:

  #if defined(OFP_DEBUG)
      // Additional debug code..
  #endif

Signed-off-by: Brian Brooks <brian.brooks@arm.com>
Reviewed-by: Jose Pekkarinen <jose.pekkarinen@nokia.com>