Fix timer double-free bug
If ofp_timer_start() is called, with a small timeout period, an odp_timer_t
will be allocated and used as the timer. This timer handle is returned to
the calling code which is also responsible for cancelling the timer.

The current implementation additionally stores this timer handle in a
global variable. This variable always holds the most recently allocated
ODP timer. There is also cleanup code which cancels this timer handle,
which causes a double-free scenario if the module managing the timer has
already canceled the timer.

This patch removes the global variable which stores a handle to the
most recently allocated ODP timer. It is not needed and causes extra
complexity as described above.

Signed-off-by: Brian Brooks <brian.brooks@arm.com>
Reviewed-by: Jose Pekkarinen <jose.pekkarinen@nokia.com>