Add unit tests for 'ofpi_in_var.h'
Add unit tests to cover functionality in 'ofpi_in_var.h'. With this change
the unit test coverage of the file in question increases from 0 to 100%
though there is still some functionality that these tests don't cover (like
are the locks properly aqcuired and released during an operation). Valgrind
was used during test creation to verify no leaks are introduced.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Signed-off-by: Sorin Vultureanu <sorin.vultureanu@enea.com>
Reviewed-and-tested-by: Sorin Vultureanu <sorin.vultureanu@enea.com>