Add unit tests for 'ofpi_tcp_seq.h'
Add unit tests to cover functionality in 'ofpi_tcp_seq.h'. The tests
assume that there will be at most 1 tick/1 ms.

The test may crash with SIGSEGV during test teardown phase. The timer
in ODP side is using threads (SIGEV_THREAD) to handle timeouts and is
passing a pointer to the timeout handler that may or may not be valid
when the timer triggers. The probability of the crash increases when
the test is ran under Valgrind.

Valgrind was used during test creation, the only leak found is related
POSIX 'timer_create' function that ODP uses to create an interval timer:
https://bugs.launchpad.net/ubuntu/+source/glibc/+bug/483594

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>