Refactor rtl search function
The original rtl search function used a fixed sized match table that,
with enough matches, could overflow. The function was anyways always
returning the latest found match so the table structure was not really
used.

The code to traverse the tree was found to be duplicated so a new
function and unit test are added. The duplicate code is removed and
the new function taken into use.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>