Add unit tests for adding and removing rules
Add unit tests to cover adding and removing of rules. Valgrind was
used during test creation to verify no leaks are possible.

Unit test for a case where the rules are full was left out as filling
up the rule table takes far too long (on one environment it took over
30s). This is because the whole table is first searched for an existing
rule and if none is found the table is searched again for the first
unused rule which in the worst case will again go through the whole
table.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>