Improve timeout handling of ofp_select
Posix select will return immediately if both fields of the timeval
structure are zero. To make the ofp_select behave in a similar way
a new check is added to skip the sleeper when both fields in ofp_timeval
structure are zero. A new unit test is also added to verify this.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-by: Bogdan Pricope <Bogdan.Pricope@enea.com>