fpm_burstmode: add lockess mq operation using input hashing
Use lockless packet receive/transmit by utilizing  multiple input/output
queues and input hashing. This enables the application throughput to scale
when the number of worker cores increases.

Processing threads are now pinned to the CPU cores using ODP default
worker mask. Threads are pinned to cores starting from the largest core id.

The core count argument ('-c') is now the actual number of workers, not
'num_workers + 1' as it used to be. The application can also be run with
only a single worker core.

Signed-off-by: Matias Elo <matias.elo@nokia.com>
Signed-off-by: Petri Savolainen <petri.savolainen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>