Add initial draft version of ofp_epoll_create
This is a very early draft of ofp_epoll_create implementation. The
actual socket creation for epoll usage is left unimplemented and is
not even needed for unit testing purposes. The mock socket creation
function used in the unit testing shows roughly what kind of changes
should be done to the actual function when support for epoll sockets
is added.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-and-tested-by: Brian Brooks <brian.brooks@linaro.org>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>