Integrate epoll into rest of the OFP
Epoll instance creation will now create a new socket, set it's type to
epoll and initialize the epoll set. OFP's socket creation has not yet
been updated to support epoll sockets so the creation here is a
temporary workaround. Socket getter, readability check and the blocking
sleep are also connected to the rest of the ofp.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-and-tested-by: Brian Brooks <brian.brooks@linaro.org>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>