C undefined behavior fixes
When both E1 and E2 in E1 << E2 have signed types, if E1 is a negative value
the behavior is undefined (S6.5.7). To prevent this, fixup cases where the
decimal literal for E1 fits into a signed integer type by using a hexadecimal
literal that only fits into an unsigned type.

Signed-off-by: Brian Brooks <brian.brooks@linaro.org>
Reviewed-by: Antero Nousiainen <antero.nousiainen@nokia.com>