Fix compilation issue with redefinition of a struct
The struct iphdr is defined in netinet/ip.h and in linux/ip.h. The
former was included by ofp_netlink.c and the latter by
linux/if_tunnel.h, which in turn was included by ofp_netlink.c. To be
backwards compatible ofp_netlink.c is modified to include linux/ip.h.

    In file included from /usr/include/linux/if_tunnel.h:6:0,
                 from ofp_netlink.c:16:
    /usr/include/linux/ip.h:85:8: error: redefinition of ‘struct iphdr’
     struct iphdr {
            ^~~~~
    In file included from ofp_netlink.c:12:0:
    /usr/include/netinet/ip.h:44:8: note: originally defined here
     struct iphdr
            ^~~~~

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>