Make epoll set size configurable
As a quick fix for the increased memory consumption, introduced by the
epoll implementation, the epoll set size is made configurable with
EPOLL_SET_SIZE definition found at api/ofp_config.h.

After this change it becomes relatively easy to add file descriptors to
epoll past the size limit. To make sure this does not cause a crash a
new unit test (along with the needed changes to the production code) is
added to see this scenario is properly handled.

Signed-off-by: Antero Nousiainen <antero.nousiainen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>