Added support for isolation into a schedule group
A new parameter in the global params allows to isolate OFP's queues in a
schedule group. If a different group is specified, any thread trying to
odp_schedule() OFP's events should be in that schedule group.

By default, OFP's behavior is backwards-compatible, all threads are
assigned to the main schedule group.

Signed-off-by: Oriol Arcas <oriol@starflownetworks.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>