tcp: Fix crash with sigevent
Fix a crash generated when processing TCP packets through sigevent
mechanism. Basically, it is caused by the fact that same memory zone
is used (unprotected) by all accepted TCP sockets.

Each socket contains a local struct ofp_sigevent. It is initialized
when sigevent functionality is enabled on a sockets (TCP listening
socket) and is copied on each TCP accepted socket. That means, for
each TCP accepted socket (+ listening socket),
so->so_sigevent.ofp_sigev_value.sival_ptr is pointing to the same
memory area. This is fine for a read-only operations but not fine for
write operations.

Fix is to provide to sigevent callback a locally stored (and
independent) sigval.

Fixes: #112 OFP TCP crash with sigevent

Signed-off-by: Bogdan Pricope <bogdan.pricope@linaro.org>
Reviewed-by: Jere Leppänen <jere.leppanen@nokia.com>