Fix erroneous prefix comparison in ofp_rt_rule_find_prefix_match
The Below two bugs exist in the same line so fix in one patch

Fixes: #114 ofp_rt_rule_find_prefix_match mix use of big-endian data
and cpu-endian data

Fix of Bug 114: store cpu-endian addr instead of big-endian addr in
rule table. Compoare the cpu-endian addr in rule with pass-in
cpu-endian addr in ofp_rt_rule_find_prefix_match so the issue is
solved. In addition the benefit of storing cpu-endian addr instead of
big-endian in rule is that it avoids ntoh of every rule to compare
with the pass-in cpu-endian addr when searching in rule table.

Fixes: #115 ofp_rt_rule_find_prefix_match use incorrect masklen to
search in rule table

Fix of Bug 115: use the shorter masklen when doing the comparison of
most significant bits. See But 115 description for details.

Signed-off-by: Dapeng Liu <dapeng.liu@linaro.org>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>