Fix the issue that a deleted mtrie route get re-inserted
Fixes: #116 ofp_rtl_remove will re-insert the route it just
removed

Analysis and Fix of Bug 116:
ofp_rtl_remove re-insert the mtrie entry that it just removed as below
steps explained.
step-1) lookup the rule table to find the rule corresponding to the
route that it is about to delete. If rule is not found then return
because the route doen't exist. If rule is found then proceed to
step-2
step-2) delete mtrie entry from mtrie
step-3) lookup the rule table again. It looks for a rule which has
shorter or equal masklen and the same prefix with the deleted route
i.e. a covering rule. If it finds such a rule then it insert the route
corresponding to this rule in mtrie.
*Here the error happens* It actually finds exactly the same rule as
step-1 which corresponds to the deleted route. So it reinserts route
it just deleted in mtrie!
step-4) when ofp_rtl_remove return to del_route. del_route delete the
rule found in step-1. So the bug has below two serious consequence.
  1) The route supposed to be delete is not deleted
  2) mtrie has the route while rule table has not corresponding rule,
  which leaves a inconsistent state.

*What is the expected behaviour* Use two examples to
elaborate. Example-1 it deleted 10.10.128.0/24 in mtrie, then it finds
there is 10.10.128.0/20 in rule table so it insert 10.10.128.0/20 in
mtrie to fill the 'hole' caused by digging 10.10.128.0/24. Example-2
it deletd 10.10.128.0/24 in mtrie, then it finds a different
10.10.128.0/24 which was overriden by current being deleted one, so it
inserts the another 10.10.128.0/24 in mtrie.

*The missing logic is* that when the code walks through rule table for
a covering rule it needs to make sure the rule being examined is not
the exact rule being deleted. It's not as easy as compare the masklen
because case like example-2 exists...

*Fix* Use the index into the rule table as identifier of a rule. With
this identifier we can exclude the deleted rule when searching in the
rule table.

*Why the issue doesn't emerge before*: The issue is hidden by
bug114. In bug114 ofp_rt_rule_find_prefix_match compares a big-endian
addr with host-endian addr when searching the rule table. Therefore
ofp_rt_rule_find_prefix_match in most case would find nothing, neither
a true covering rule or the rule corresponding to the deleted
route. Now with bug114 fixed in another patch the ofp_rtl_remove issue
is exposed.

*two test cases Modified and two new test cases Added*:
1) Remove route and not re-inserted - M
2) Remove route with second level mask and not re-inserted - M
3) Remove route reinserted when there exists a different rule covering
the rule being deleted - A
4)Remove route with second level mask reinserted when there exists a
different rule covering the rule being deleted - A

Signed-off-by: Dapeng Liu <dapeng.liu@linaro.org>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>