Fix Undefined Behaviour about left bit shift operation
Fixes: #119 - Left bit shift result in UB

For E1 << E2, according to C programming language, the result is
undefined if the right operand is negative, or greater than or equal
to the number of bits in the left expression's type. For example on my
testbed 0xFFFFFFFF << (32 - masklen) result in 0xFFFFFFFF instead of 0
when masklen is 0, which is clearly wrong. This patch corrects all the
flawy left bit shift about mask as right hand operand.

Signed-off-by: Dapeng Liu <dapeng.liu@linaro.org>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>