example: add new tcpperf example application
Add new example application for testing ofp tcp socket throughput. The
application operates in either server or client mode. In server mode the
application works as a sink for incoming tcp data. In client mode it tries
to send data to the target socket as fast as possible. In both cases the
maximum transmission rates are measured. The application uses a single
socket for communication.

In addition to running tcpperf on both ends of the link one can also use
the standard linux iperf application (tcpperf client -> iperf server or
iperf client -> tcpperf server).

Signed-off-by: Matias Elo <matias.elo@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>