tcp: Fix fast retransmit on three duplicate acks.
Admittedly this is a hack, but it seems to work, or at least improve
the behavior. The idea is to make sure that after receiving the third
dupack, the necessary conditions hold so that we retransmit
immediately.

Fast retransmits have probably gotten broken when some of the tcp code
has been disabled, especially some of the congestion control related
code.

This patch fixes one specific fast retransmit problem, others may
still remain.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>