Increase packet pool size in the WEBSERVER configuration.
Since socket buffers use packets as storage, a socket using
application may end up using quite a few packets. The current number
of packets in the pool corresponds to 10240 / 64 = 160 full socket
buffers, which is a small number for a web server.

Increase the number of packets in the pool to the next order of
magnitude, which is 100000. This requires more physically contiguous
hugepages (on odp-dpdk), but if that is a problem to the user, it's a
much clearer problem than running out of packets in the pool, a
situation which often results in the system not being able to send or
receive anything.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>