Set the language standard to C99.
To avoid surprises caused by different versions of gcc defaulting to
different language standards, set it explicitly to C99. Many people
may still be using older gcc versions with incomplete C11 support, so
C99 is probably the most reasonable choice.

The _DEFAULT_SOURCE feature macro is needed to bring in various
features, which are used in many places in the code, and are *not* a
part of C99.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>