Call default settings of parameters before OFP init global
In the examples, before starting the OFP, the global parameters were set to 0.
Instead of this, the newly added ofp_init_global_param function should be used
to set the structure to its default value. Every example that uses ofp_init_global
function to start the OFP process is changed to use this setting.

Signed-off-by: Iulia Florea <iulia.florea@enea.com>
Reviewed-by: Jere Leppänen <jere.leppanen@nokia.com>