icmp: Don't reflect IP options.
When sending ICMP using icmp_reflect(), the received packet is used as
a base. IP options are retained, but IP header length is nevertheless
reset to 20 bytes. This causes the following problem:

$ echo "hello" | socat - UDP-DATAGRAM:11.0.0.10:65000,ip-options=x88041111
12:02:45.498600 IP (tos 0x0, ttl 64, id 21713, offset 0, flags [DF], proto UDP (17), length 38, options (unknown 136))
    11.0.0.20.43508 > 11.0.0.10.65000: [udp sum ok] UDP, length 6
13:17:28.405560 IP (tos 0x0, ttl 64, id 8, offset 0, flags [DF], proto ICMP (1), length 64)
    11.0.0.10 > 11.0.0.20: ICMP type-#136, length 44

Type 136 = 0x88, which is actually the number of the first IP option
in the triggering packet.

Since we don't have any IP option processing, the fix is to remove all
IP options in icmp_reflect().

Fixes: #179 ICMP packet is malformed if triggering packet contains IP options.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>