Use ordered instead of atomic queues in fpm to improve throughput.
Use of atomic queues limit parallelism in fpm when forwarding one or
few traffic flows. The effect is the higher the heavier the processing
is per packet.

Use ordered input queues to allow parallel processing of packets that
belong to the same flow and use queued output to get the packets to
the original order in output.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>