Increase the default number of packet buffers to 10240 in all cases.
The current default number of packet buffers is not sufficient when using
DPDK based packet IO with the linux generic ODP. There is prepocessor
code for increasing the default when ODP-DPDK is used but that obviously
does not help when Linux generic ODP is used.

Get rid of the #ifdefs and make the number of packet buffers always the
higher one of the two to make it more likely that OFP works out of the box.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>