api: Use the term "CLI file" instead of "configuration file".
Don't use the term "configuration file". Since the file contains CLI
commands, call it a "CLI file". This is done in order to make room for
configuration file support that will be implemented later.

The only changes to the API is documentation and a parameter name, so
there's no effect to applications.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>