Append a freed rtl node at the end of the list of free available nodes.
This commit is to fix Github Issue 120. The original code added free'd nodes
to the head of the free list, and this change adds them to the tail. This
prevents a recently free'd node to be reused immediately. Thus keeping the
data of the node available for longer time in case some other worker thread
is using it.
The orginal test code of mtrie has also been changed as it was based on an
assumption that a recently free'd node is added to the head of the free list.

This assumption had following code implications which has been changed
hereafter.

1. The next pointer of recently freed node was assumed to be NOT NULL
as it was added to the head of the free list. This is no more true.
2. The node most recently freed is used to allocate a new rtl node.
This also is no more true.

Fixes: #120 Nexthops may be used after freed and while being modified

Signed-off-by: Sirshak Das <sirshak.das@arm.com>
Reviewed-by: Brian Brooks <brian.brooks@arm.com>
Reviewed-by: Honnappa Nagarahalli <honnappa.nagarahalli@arm.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>