arp: Fix check for ARP entry existence when saving packets.
Apparently the check for MAC address under ARP_SANITY_CHECK in
ofp_arp_save_ipv4_pkt() was supposed to check the last byte of the
address, but there's an off by one error. Since the MAC address is
stored as an uint64_t, fix this by checking the whole field.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>