Make the number of 16 bit MTRIE nodes equal to the number of VRFs.
Each VRF has its own MTRIE, so the number of MTRIE root nodes (the 16
bit nodes) is equal to the number of VRFs. Remove the
mtrie.table16_nodes initialization parameter, and instead allocate as
many nodes as there are VRFs.

A 16 bit MTRIE node takes 4MB of space, so it's necessary to reduce
the default number of VRFs. Since the number of VRFs is configurable,
it's logical to set it to 1 by default.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>