Fix l4_offset of incoming packets
Fixes: #189 Outgoing TCP RST contains bad TCP header fields.

Update the l4_offset of an incoming packet within ofp_tcp_input() after the
L2 header is removed.

This patch fixes a problem that occurs when an incoming TCP packet triggers
a RST. The ODP packet header of the resulting RST contains the wrong
l4_offset, which causes the application to access the TCP header fields at
the wrong offset.

The root cause of the problem is that the function ofp_tcp_input() removes
the L2 header from the beginning of the received packet but only updates
the l2_offset and l3_offset fields within the ODP packet header, and not
the l4_offset.

If ofp_tcp_input later finds any issues with this incoming packet which
make it necessary for the packet to be dropped and a RST to be sent back,
it will call ofp_tcp_respond (via tcp_dropwithreset) to construct a RST
response using the ODP buffer of the incoming packet. The ODP packet header
of the newly created RST will inherit the wrong l4_offset and will cause
the TCP header fields to be accessed at the wrong offset.

The patch below fixes the problem by setting the correct l4_offset for the
incoming packet within ofp_tcp_input().

Signed-off-by: Janos Villanyi <a74esev@gmail.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>