Allow ofp_stop_processing() calls in uninitialized contexts.
ofp_stop_processing() may be called by a signal handler running in the
context of a thread or process that has not done OFP initialization.
Check the shm pointer before using it.

This is a workaround for an ODP problem and avoids annoying errors when
stopping fpm. See https://bugs.linaro.org/show_bug.cgi?id=3690

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>