test: Fix frame length checks by fixing test frames.
Test frames contained extra bytes after the IP packet and expected
the same bytes be present in result packets after OFP processing.
The fix in commit c226ac7 exposed this problem in the test code and
caused the tests to fail.

Fix the problem by removing the extra bytes from the end of the test
frames so that the comparisons of the result packets with the test
frames work as supposed to.

Signed-off-by: Janne Peltonen <janne.peltonen@nokia.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>