avl: Use ofp_brlock instead of odp_rwlock.
When multiple threads are read locking the same rwlock frequently,
e.g. for each packet, the locking can become quite slow. This happens
for example when multiple threads are receiving packets to VLANs on
the same physical interface, because the VLANs are stored in an AVL
tree, which uses an odp_rwlock. Use an ofp_brlock instead.

Signed-off-by: Jere Leppänen <jere.leppanen@nokia.com>
Reviewed-by: Valentin Radulescu <valentin.radulescu@enea.com>