Fixes: #204 TCP src and dst ports mixed up in IPv6 RST
Swap the src and dst ports in the outgoing RST for IPv6 as well, as is
already done for IPv4.

The ofp_tcp_respond() function reuses the mbuf of an incoming bad packet to
generate a RST response. This of course means that both the IP addresses
and the TCP ports need to be swapped within the buffer for the new packet.
Currently the IP addresses are correctly swapped, but the TCP ports are
only swapped in the IPv4 case.

The patch below fixes this by rearranging the code so that the ports are
swapped both for IPv4 and for IPv6.

Signed-off-by: Janos Villanyi <a74esev@gmail.com>
Reviewed-by: Valentin Radulescu <valentin.radulescu@enea.com>